пересмотрен подход к проектированию телеграмм бота
